Added tools for testing correctness of derivatives
These are only compatible with autograd currently, and the riemannian
hessians will not be efficient enough to use for optimisation. However
in future this will be possible.

These methods should be used for testing the manifold’s egrad2rgrad and
ehess2rhess methods. Note though, that they only work for sub manifolds
of Euclidean space (and possibly quotient manifolds like the Grassman —
need to check this).

rhess could be adapted to work for general manifolds.