Update MoG.
Switch from theano to autograd for cost definition. This seemed much faster
in my testing. Also vectorise everything and make it a lot easier to add
another cluster if you want to. Switch to steepest descent since this
seemed to work better for very large n (and still very fast).

Signed-off-by: Jamie Townsend <jamiehntownsend@gmail.com>