Correct parameterization FixedRankEmbedded
Assume that the cost function is defined as a function of the full low rank
matrix (rather than its singular value decomposition). This is inefficient
but necessary for now. This required defining a custom class _Point for points
on the FixedRankEmbedded manifold and in turn this has required changing the
autograd backend to automatically cast to numpy.ndarray before passing anything
to numpy/autograd.

This seems quite inelegant and hopefully can be made nicer in the long term.

Signed-off-by: Jamie Townsend <jamiehntownsend@gmail.com>