Reorganize backend logic (theano only so far)
Cost functions, gradients and Hessian-vector products must now always be
defined to accept arguments in an unpacked form. This makes it more
explicit how points on manifolds are represented. In order to work with
product manifolds, function decorators require a variable number of
arguments to define the respective argument grouping.

This currently breaks most if not all backend related unit tests.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>