Move version info to pymanopt/_version.py
In order to build the sphinx doc, we need to get the current version of
pymanopt. To that end we use the `runpy` module to fetch the info from
the source directory. However, if we keep the info field in the toplevel
package's `__init__.py`, sphinx tries to import all of pymanopt on
build, which requires us to either set the `PYTHONPATH` variable or
install pymanopt after successful builds on travis. To avoid this, we
make the version info file independently importable.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>