Refactor 'Manifold' base class
This commit separates methods in the 'Manifold' base class into abstract
methods, which must be implemented by valid subclasses, and those which
are only required by certain solvers. If a subclass does not provide one
of those optional methods, we now raise a generic but informative
NotImplementedError.

The class constructor now also requires a 'name' and 'dimension'
argument, which are used in the generic implementations of the classes
'__str__' method and 'dim' property so that subclasses don't have to
provide this method/property manually.

Embedded submanifolds of Euclidean spaces offer particularly convenient
ways to convert Euclidean gradients and Hessian-vector products to their
Riemannian counterparts by means of tangent space projectors and
Weingarten maps, respectively. To ease implementations, we now provide a
'EuclideanEmbeddedSubmanifold' class, which, in addition to the required
abstract 'Manifold' methods, only requires descendants to implement the
'weingarten' method to support 'ehess2rhess'.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>