sphere.py: refactor w.r.t. new base classes
This commit adapts the sphere manifold classes to the new base class
interface. Most notably, this breaks API with the existing
'SphereSubspaceIntersection' and 'SphereSubspaceComplementIntersection'
classes, whose constructors previously required the dimension of the
ambient space. The new constructors now only accept a single matrix
whose columns span the respective subspace. The dimension of the ambient
space is simply chosen as the length of the vectors in the spanning set.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>