_pytorch.py: handle ndarrays with negative strides
PyTorch's 'from_numpy' function currently does not support converting
numpy arrays with negative strides. Such arrays arise for instance when
reversing the order of elements via 'array[::-1]', which creates a view
of the array with a negative stride rather than copying the array. Since
torch does not support this, we have to detect negative strides in the
autodiff backend and create a copy of the array if necessary.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>