Retrieve signature hint from manifolds
This is the first step in shifting the responsibility to obtain
signature hints of cost functions, gradients and Hessian-vector products
from the user to the toolbox.

As a proof-of-concept, this commit only deals with the autograd backend,
whose implementation is now significantly simpler as appropriate wrapping
of the respective function is now taken care of by the 'Problem' class.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>