Require manifolds to define 'tangent' method
For the upcoming 'check_gradient' functionality, manifolds need to
implement the 'tangent' method to re-tangentialize tangent vectors which
might violate the tangent space condition due to numerical error
accumulation. For most manifolds, this simply corresponds to a
projection of the tangent vector on the tangent space. However, since
this behavior cannot be inferred from a property (such as a manifold
being an embedded submanifold of a Euclidean space), we require
manifolds to be explicit about the choice for 'tangent' rather than
providing a default implementation which simply calls 'proj'.

Moreover, this change makes sure that we never simply assign methods
such as 'retr' or 'egrad2rgrad' on class level as this prevents that the
methods inherit their corresponding docstrings from the base class.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>