Remove function signature check
If decorated functions accepted positional keyword arguments, we
previously verified that the signature matches the manifold's point
layout. However, hvps require twice the number of arguments compared to
regular cost functions. Since we do not know whether we are decorating a
cost function or an hvp, we have to forgo any attempts to verify the
function signature beforehand.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>