Rename 'hvp' to 'hessian' for clarity
While the term 'hvp' is more accurate to describe what we are generally
referring to, the terminology is still slightly off and may lead to more
confusion. For that reason, we rename any functions referring to 'hvp'
in their name to use the more common term 'hessian' where necessary.
When in doubt, the docstrings should be sufficient to clarify what
e.g. 'euclidean_hessian' really represents.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>