Convert tangent vectors to correct format
For Euclidean Hessian operators it is natural to expect tangent vectors
to be represented in the ambient space. Optimizers, however, pass
tangent vectors in the internal representation of the manifold when
calling `riemannian_hessian`. A typical example are tangent vectors of
manifolds which are also Lie groups, in which case tangent vectors can
be conveniently represented in the Lie algebra.

Since `riemannian_hessian` usually calls `euclidean_hessian` under the
hood, this leads to a discrepancy. Most crucially, this affects Hessian
operators which are automatically generated by our autodiff backends.
This commit fixes the previous behavior by adding a call to
`manifold.embedding` before tangent vectors are passed to Euclidean
Hessian operators, no matter if the operator was user-defined are
generated automatically.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>